+++ lib/src/display/segment_display.dart
@@ -15,12 +15,12 @@ import 'package:segment_display/src/segment_display_painter.dart';
 /// and implement [createSingleCharacter] method.
 abstract class SegmentDisplay extends StatelessWidget {
   /// Characters to be displayed on display.
-  final String text;
+  final String value;
 
   /// Size of characters on display.
   ///
-  /// NOTE: [SegmentStyle.segmentBaseSize] * [textSize] = [segmentSize]
-  final double textSize;
+  /// NOTE: [SegmentStyle.segmentBaseSize] * [size] = [segmentSize]
+  final double size;
 
   /// Style for segments which defines shape, color or size for segments.
   ///
@@ -29,16 +29,16 @@ abstract class SegmentDisplay extends StatelessWidget {
 
   /// Number of characters to display, used when building display.
   ///
-  /// The [text] length will be used when no count is provided.
+  /// The [value] length will be used when no count is provided.
   ///
-  /// If [characterCount] > [text.length] then text will be left padded.
-  /// This also means if [characterCount] < [text.lenght] then
+  /// If [characterCount] > [value.length] then text will be left padded.
+  /// This also means if [characterCount] < [value.length] then
   /// only last X characters will be displayed (based on [characterCount]).
   ///
   /// For example:
-  ///  * [characterCount] is set to 3 and [text] value to "1" -
+  ///  * [characterCount] is set to 3 and [value] value to "1" -
   ///     this will be displayed like "__1" (left padded).
-  ///  * [characterCount] is set to 1 and [text] value to "123" -
+  ///  * [characterCount] is set to 1 and [value] value to "123" -
   ///     this will be displayed like "3" (only last character).
   final int characterCount;
 
@@ -58,14 +58,14 @@ abstract class SegmentDisplay extends StatelessWidget {
 
   const SegmentDisplay({
     Key key,
-    @required this.text,
+    @required this.value,
     @required this.characterMap,
-    double textSize,
+    double size,
     SegmentStyle segmentStyle,
     this.characterCount,
     double characterSpacing,
     Color backgroundColor,
-  })  : textSize = textSize ?? 10.0,
+  })  : size = size ?? 10.0,
         segmentStyle = segmentStyle ?? const DefaultSegmentStyle(),
         characterSpacing = characterSpacing ?? 7.0,
         backgroundColor = backgroundColor ?? const Color(0xff000000),
@@ -75,26 +75,25 @@ abstract class SegmentDisplay extends StatelessWidget {
   ///
   /// * [Size.width] represents 'thickness' of segment
   /// * [Size.height] represents 'length' of segment
-  Size get segmentSize => segmentStyle.segmentBaseSize * textSize;
+  Size get segmentSize => segmentStyle.segmentBaseSize * size;
 
   @override
   Widget build(BuildContext context) {
-    final charCount = characterCount ?? text.length;
-    final size = computeSize(charCount);
+    final _displaySize = computeSize();
 
     return Semantics(
       label: "Segment display",
-      value: text,
+      value: value,
       textDirection: TextDirection.ltr,
       child: Container(
         color: backgroundColor,
         child: SizedBox(
-          width: size.width,
-          height: size.height,
+          width: _displaySize.width,
+          height: _displaySize.height,
           child: CustomPaint(
-            size: size,
+            size: _displaySize,
             painter: SegmentDisplayPainter(
-              segments: createDisplaySegments(charCount, size),
+              segments: createDisplaySegments(),
               enabledColor: segmentStyle.enabledColor,
               disabledColor: segmentStyle.disabledColor,
             ),
@@ -108,18 +107,30 @@ abstract class SegmentDisplay extends StatelessWidget {
   bool canDisplay(String character) => characterMap.containsKey(character);
 
   /// Creates whole display - all characters (all segments)
-  List<Segment> createDisplaySegments(int charCount, Size size) {
+  List<Segment> createDisplaySegments() {
+    final charCount = characterCount ?? value.length;
     final segments = <Segment>[];
 
+    double indent = 0;
+
     // Add left padding when characterCount > text.length
-    for (var i = text.length - charCount; i < text.length; i++) {
+    for (var i = value.length - charCount; i < value.length; i++) {
       var char = "";
-      if (i >= 0 && i < text.length) {
-        char = text[i];
+      if (i >= 0 && i < value.length) {
+        char = value[i];
       }
 
-      final characterSegments =
-          createSingleCharacter(i - (text.length - charCount));
+      // handle dividers rendering
+      if (CharacterSegmentMap.dividerCharacters.values.contains(char)) {
+        final segment = createDividerCharacter(char, indent);
+        if (segment != null) {
+          indent += segmentSize.width + characterSpacing;
+          segments.add(segment);
+        }
+        continue;
+      }
+
+      final characterSegments = createSingleCharacter(indent);
 
       // when character can be displayed, enable given segments
       if (char.isNotEmpty && canDisplay(char)) {
@@ -131,6 +142,7 @@ abstract class SegmentDisplay extends StatelessWidget {
         }
       }
 
+      indent += (2 * segmentSize.width + segmentSize.height) + characterSpacing;
       segments.addAll(characterSegments);
     }
 
@@ -138,14 +150,34 @@ abstract class SegmentDisplay extends StatelessWidget {
   }
 
   /// Returns current [Size] of whole display.
-  Size computeSize(int charCount) {
-    final width = charCount * (2 * segmentSize.width + segmentSize.height);
-    final widthOffset = characterSpacing * (charCount - 1);
+  Size computeSize() {
+    final dividers = CharacterSegmentMap.dividerCharacters.values;
+
+    final charCount = characterCount ??
+        value.split("").where((c) => !dividers.contains(c)).length;
+    final dividerCharCount = value.split("").where(dividers.contains).length;
+
+    final charsWidth = charCount * (2 * segmentSize.width + segmentSize.height);
+    final dividersWidth = dividerCharCount * segmentSize.width;
+    final width = charsWidth + dividersWidth;
+    final widthOffset = characterSpacing * (charCount - 1) +
+        (characterSpacing * dividerCharCount);
     final height = (2 * segmentSize.height) + (3 * segmentSize.width);
 
     return Size(width + widthOffset, height);
   }
 
-  /// Creates segments to display a single character.
-  List<Segment> createSingleCharacter(int charIndex);
+  /// Creates segments to display a (single) character.
+  List<Segment> createSingleCharacter(double indent);
+
+  /// Creates special segment to display dividers like decimal point or colon.
+  Segment createDividerCharacter(String character, double indent) {
+    if (character == CharacterSegmentMap.dividerCharacters["decimalPoint"]) {
+      return Segment.decimalPoint(segmentStyle, segmentSize, indent);
+    } else if (character == CharacterSegmentMap.dividerCharacters["colon"]) {
+      return Segment.colon(segmentStyle, segmentSize, indent);
+    } else {
+      return null;
+    }
+  }
 }